backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pypdf==6.1.3
pdfplumber==0.9.0
numpy==1.26.4
pandas==2.2.3
openpyxl==3.1.2
xlrd==2.0.2
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
waitress==3.0.0
requests==2.31.0
psutil==5.9.6
flasgger==0.9.7.1
pyyaml==6.0.1
jsonschema==4.20.0
bleach==6.0.0
boto3==1.34.0
